Moved OpenIdConfigKeyEnum out of seedwork files
