+++ data-access/src/routes/storage-queue/init/queue-context-builder.ts
@@ -1,7 +1,6 @@
 import { InvocationContext } from "@azure/functions";
 import { InfrastructureServices } from '../../../app/infrastructure-services';
-import { AppContext, AppContextBuilder, VerifiedUser } from '../../../app/init/app-context-builder';
-import { OpenIdConfigKeyEnum } from "../../../../seedwork/auth-seedwork-oidc/portal-token-validation";
+import { AppContext, AppContextBuilder, OpenIdConfigKeyEnum, VerifiedUser } from '../../../app/init/app-context-builder';
 
 
 export interface QueueContext extends AppContext{
